+++ usbd-storage/src/subclass/scsi.rs
@@ -6,7 +6,7 @@ use core::fmt::Debug;
 use num_enum::TryFromPrimitive;
 use usb_device::bus::InterfaceNumber;
 use usb_device::bus::UsbBus;
-use usb_device::class::{ControlIn, UsbClass};
+use usb_device::class::{ControlIn, ControlOut, UsbClass};
 use usb_device::descriptor::DescriptorWriter;
 #[cfg(feature = "bbb")]
 use {
@@ -333,6 +333,10 @@ where
 
     fn control_in(&mut self, xfer: ControlIn<Bus>) {
         self.transport.control_in(xfer)
+    }
+
+    fn control_out(&mut self, xfer: ControlOut<Bus>) {
+        self.transport.control_out(xfer)
     }
 }