+++ usbd-storage/src/subclass/ufi.rs
@@ -4,7 +4,7 @@ use crate::CLASS_MASS_STORAGE;
 use crate::transport::Transport;
 use usb_device::bus::InterfaceNumber;
 use usb_device::bus::UsbBus;
-use usb_device::class::{ControlIn, UsbClass};
+use usb_device::class::{ControlIn, ControlOut, UsbClass};
 use usb_device::descriptor::DescriptorWriter;
 #[cfg(feature = "bbb")]
 use {
@@ -330,5 +330,9 @@ where
 
     fn control_in(&mut self, xfer: ControlIn<Bus>) {
         self.transport.control_in(xfer)
+    }
+
+    fn control_out(&mut self, xfer: ControlOut<Bus>) {
+        self.transport.control_out(xfer)
     }
 }